core-simulator/cli/requirements.txt
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
pyaml==25.7.0
Pygments==2.19.2
PyYAML==6.0.2
rich==14.1.0
shellingham==1.5.4
sniffio==1.3.1
typer==0.16.0
typing_extensions==4.14.0
